feat: 지원 관리 관련 상수 분리
